refactor(frontend): collapse 5 hljs utils into 2 along the worker boundary
The hljs plumbing had grown to 5 separate files (HighlightLoader,
HighlightContext, HljsProvider, hljsLanguages, useHLJS). Merge them
into 2, keeping the only architectural split that actually matters —
the worker-SSR safety boundary.

  highlight.ts        worker-safe: highlightHTML, contexts + consumer
                      hooks, type re-export. No runtime hljs import,
                      so any component (including ones reached by the
                      worker SSR entries) can pull it.

  highlight-client.tsx client-only: the import.meta.glob over the hljs
                      language files, alias resolution, the useHLJS
                      hook, the HljsProvider component. Pulls hljs at
                      runtime — only client entries may import.

Touches every caller: DisplayPasteView / CodeEditor switch to a single
`../utils/highlight.js` import; render/{index,display,markdown} point
at `../../utils/highlight-client.js`. No behaviour change, no test or
bundle diff (still 977KB raw / 198KB gzip worker, 0 hljs references).
