+++ frontend/pages/DisplayPasteView.tsx
@@ -4,8 +4,7 @@ import { DarkModeToggle, useDarkModeSelection } from "../components/DarkModeTogg
 import { DownloadIcon, HomeIcon } from "../components/icons.js"
 import { CopyWidget } from "../components/CopyWidget.js"
 import { tst } from "../utils/overrides.js"
-import { highlightHTML } from "../utils/HighlightLoader.js"
-import { useHljsForLang } from "../utils/HighlightContext.js"
+import { highlightHTML, useHljsForLang } from "../utils/highlight.js"
 import { formatSize } from "../utils/utils.js"
 
 interface PendingInfo {
