+++ frontend/pages/render/index.tsx
@@ -1,7 +1,7 @@
 import { hydrateRoot, createRoot } from "react-dom/client"
 import React from "react"
 import { PasteBin } from "../PasteBin.js"
-import { HljsProvider } from "../../utils/HljsProvider.js"
+import { HljsProvider } from "../../utils/highlight-client.js"
 
 const rootElement = document.getElementById("root")!
 const config = __WRANGLER_CONFIG__
